requirements.txt
pipecat-ai[deepgram,openai,elevenlabs,silero]
python-dotenv
fastapi
uvicorn
aiohttp
websockets